requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
supabase==0.5.8
python-dotenv==0.19.1
schedule==1.1.0
pytz==2021.3